折りたたみのJSを共通化
